[4/x] pass key to menu content
